chore: move getFirstResolvingPromise to utils
