+++ src/core/common/utils/index.ts
@@ -15,3 +15,4 @@ export { default as pureMerge } from "./pureMerge";
 export { default as startsWith } from "./startsWith";
 export { default as stringifyQuery } from "./stringifyQuery";
 export { default as roundRating } from "./roundRating";
+export { default as getFirstResolvingPromise } from "./getFirstResolvingPromise";
